This branch comes with a lot of other changes.
1. Fix group_by so it shouldn't blow up when iter_begin == iter_end
2. Now try to use <functional> whenever possible.
3. Of course, some close-to-correct implementation of solution to problem 18.

To expand on 3, the only problem we have right now is dealing with the number 0.
If there 1 zero, then we must find the case where 3 elements add up to `target`.

The "simplest" solution is probably to call N-Sum multiple times for N={3,2}.

+++ include/akarithm/problem_5329.hpp
@@ -36,7 +36,7 @@ int minSetSize(
         [&](const std::vector<int> &value)
             -> bool {
             acc += static_cast<int>(value.size());
-            return acc > arr.size() / 2;
+            return acc > static_cast<int>(arr.size() / 2);
         }));
 }
 } // namespace akarithm
